refactor(dao)!: replace UnsupportedOperationException override-stubs with a capability-based DAO interface hierarchy
Restrictions on ReadOnly/NoUpdate DAOs (and their
Crud/L/Unchecked/JoinEntityHelper
variants) were enforced by re-declaring every forbidden method as an
@Deprecated
default that throws UnsupportedOperationException (~165 throw-stubs
across 18 files,
plus diamond re-resolution boilerplate). This replaces that subtractive
model with an
additive capability model: restrictions are expressed by NOT inheriting
a capability,
so forbidden entity operations are now compile errors instead of runtime
exceptions.

BREAKING CHANGE: backward compatibility was intentionally waived.
- ReadOnly/NoUpdate DAOs no longer expose
save/insert/update/upsert/delete/batchXxx
  (and prepareCallableQuery/stmtCreator prepares) — calling them no
longer compiles.
- ReadOnlyDao no longer extends NoUpdateDao (the old subtype edge was a
Liskov
  violation); read-only restricted *L variants drop the primitive-long
get/deleteById
  convenience overloads (the boxed Long overloads still apply via
autoboxing).
- JdbcUtil.createDao is now bound `<TD extends ReadableDao>` (was `<TD
extends Dao>`);
  dao.DaoUtil is now public (still @Internal).

New capability interfaces (25):
- Condition level: ReadableDao, InsertableDao, UpdatableDao,
DeletableDao (+ Unchecked
  twins). Dao = Readable+Insertable+Updatable+Deletable (compound
upsert/batchUpsert +
  stmtCreator/callable prepares kept in Dao).
- CRUD level: ReadableCrudDao, InsertableCrudDao, UpdatableCrudDao,
DeletableCrudDao
  (+ Unchecked twins), and the Long-id markers ReadableCrudDaoL /
UncheckedReadableCrudDaoL.
- Join level: ReadableJoinEntityHelper, DeletableJoinEntityHelper,
  ReadableCrudJoinEntityHelper (+ Unchecked twins).
- Cacheable: marker for cache-eligible (read/insert-only) DAOs.

Composition (now thin bundles, all carrying 0 throw-stubs):
- NoUpdateDao = InsertableDao + Cacheable;  ReadOnlyDao = ReadableDao +
Cacheable
- CrudDao = Dao + the four *CrudDao caps;  NoUpdateCrudDao = NoUpdateDao
+ ReadableCrudDao
  + InsertableCrudDao;  ReadOnlyCrudDao = ReadOnlyDao + ReadableCrudDao
- JoinEntityHelper = Readable + Deletable;  ReadOnlyJoinEntityHelper =
Readable
- The hierarchy is sealed/non-sealed with explicit `permits` clauses.

DaoImpl / JdbcUtil / DaoUtil:
- Cache gating now keys off the Cacheable marker (3 sites) instead of
the NoUpdateDao
  type, so caching stays enabled on ReadOnly DAOs.
- The prepareQuery/prepareNamedQuery/*ForLargeResult SQL-kind check
(SELECT-only for
  read-only, SELECT+INSERT for no-update) is centralized in the DaoImpl
proxy
  (ParsedSql null -> IllegalArgumentException, bad SQL ->
UnsupportedOperationException
  with the original messages); the per-interface override methods are
removed.
- Join/CRUD/unchecked detection retyped to the Readable* roots
(ReadableDao /
  ReadableCrudDao / ReadableCrudDaoL / UncheckedReadableDao); the two
declaringClass
  proxy-dispatch guards broadened to the capability classes; proxy
invoker plumbing
  (BiFunction<…>, the proxy cast, daoPool,
getApplicableDaoForJoinEntity) retyped
  Dao -> ReadableDao since a read-only proxy no longer implements Dao.
- DaoUtil.isReadOnlyQuery / isNoUpdateQuery exposed as public.

Tests:
- New CapabilityGatingTest: Cacheable marker structure; caching
permitted on
  ReadOnly/NoUpdate but rejected on a full Dao; centralized prepare-gate
via real
  createDao proxies; capability composition graph.
- DaoUtilTest: direct coverage of the now-public
isReadOnlyQuery/isNoUpdateQuery.
- NoUpdate/ReadOnly (Dao/Crud/L + Unchecked) and JoinEntityHelper tests
updated:
  obsolete runtime-UOE assertions replaced by capability-structure
assertions and
  real-proxy gate tests (the restrictions are now compile-time /
proxy-enforced).

Behavior-preserving refactor (apart from the intended compile-time
restrictions):
mvn clean test -> 3265 tests, 0 failures/0 errors; mvn javadoc:javadoc
-> BUILD SUCCESS.

+++ src/test/java/com/landawn/abacus/jdbc/JdbcUtilTest.java
@@ -60,7 +60,6 @@ import com.landawn.abacus.jdbc.Jdbc.RowFilter;
 import com.landawn.abacus.jdbc.Jdbc.RowMapper;
 import com.landawn.abacus.jdbc.dao.CrudDao;
 import com.landawn.abacus.query.ParsedSql;
-import com.landawn.abacus.query.SqlBuilder;
 import com.landawn.abacus.util.ContinuableFuture;
 import com.landawn.abacus.util.Dataset;
 import com.landawn.abacus.util.ImmutableMap;
@@ -629,7 +628,7 @@ public class JdbcUtilTest extends TestBase {
     // still freed in the finally block.
     @Test
     public void testGetColumnValue_BlobSizeOverflow_Throws() throws SQLException {
-        when(mockBlob.length()).thenReturn((long) Integer.MAX_VALUE + 1L);
+        when(mockBlob.length()).thenReturn(Integer.MAX_VALUE + 1L);
         when(mockResultSet.getObject(1)).thenReturn(mockBlob);
 
         assertThrows(SQLException.class, () -> JdbcUtil.getColumnValue(mockResultSet, 1));
@@ -639,7 +638,7 @@ public class JdbcUtilTest extends TestBase {
     // Clob whose length exceeds Integer.MAX_VALUE trips the size guard (JdbcUtil L2272).
     @Test
     public void testGetColumnValue_ClobSizeOverflow_Throws() throws SQLException {
-        when(mockClob.length()).thenReturn((long) Integer.MAX_VALUE + 1L);
+        when(mockClob.length()).thenReturn(Integer.MAX_VALUE + 1L);
         when(mockResultSet.getObject(1)).thenReturn(mockClob);
 
         assertThrows(SQLException.class, () -> JdbcUtil.getColumnValue(mockResultSet, 1));
